Silence compiler warnings

+++ OpenSim/Analyses/InverseDynamics.cpp
@@ -169,7 +169,7 @@ constructColumnLabels()
            auto coordinates = _modelWorkingCopy->getCoordinatesInMultibodyTreeOrder();
            for (int i=0; i < _numCoordinateActuators; i++) {
               Force& force = _forceSet->get(i);
-              for(int i=0; i<coordinates.size(); ++i) {
+              for(size_t i=0u; i<coordinates.size(); ++i) {
                  const Coordinate& coord = *coordinates[i];
                  if(coord.getName()==force.getName()) {
                     if(coord.getMotionType() == Coordinate::Rotational) {
@@ -455,7 +455,7 @@ begin(SimTK::State& s )
         // Gather indices into speed set corresponding to the unconstrained degrees of freedom (for which we will set acceleration constraints)
         _accelerationIndices.setSize(0);
         auto coordinates = _modelWorkingCopy->getCoordinatesInMultibodyTreeOrder();
-        for(int i=0; i<coordinates.size(); ++i) {
+        for(size_t i=0u; i<coordinates.size(); ++i) {
             const Coordinate& coord = *coordinates[i];
             if(!coord.isConstrained(sWorkingCopy)) {
                 _accelerationIndices.append(i);
